code/data/preprocess/requirements.txt
# Data processing and scientific computing
numpy>=1.22.0
pandas>=1.5.0
scipy>=1.9.0
xarray>=2023.1.0
dask>=2023.1.0
distributed>=2023.1.0  # Dask's distributed scheduler

# Geospatial libraries
rioxarray>=0.13.0
rasterio>=1.3.0
pyproj>=3.4.0
shapely>=2.0.0
geopandas>=0.12.0

# HDF and NetCDF support
h5netcdf>=1.1.0
h5py>=3.7.0
netCDF4>=1.6.0

# Storage formats
zarr>=2.13.0
fsspec>=2022.11.0
gcsfs>=2023.1.0  # For Google Cloud Storage

# Google Cloud
google-cloud-storage>=2.6.0
google-auth>=2.15.0

# Utilities
tqdm>=4.64.0  # Progress bars
pyyaml>=6.0   # For YAML config files
python-dotenv>=0.21.0  # For loading .env files

# Logging and monitoring
logging>=0.5.1